# changed code to use []byte/string conversions.