changed code to use []byte/string conversions where applicable. I believe there is a comment within the source referencing this optimization, one of few optimizations that exist within the gc toolchain. first of probably several commits regarding conversions.

this code accomplishes exactly the same goal as the previous code. it is just faster and clearer.