@valyala 2016-01-06
@sprstnd , your approach results in performance degradation on the following server benchmarks on my laptop:

```
$ git checkout master
$ go test -bench=kServerGet.*10KClients -benchmem -benchtime=5s > old.txt
$ git checkout sprstnd-master
$ go test -bench=kServerGet.*10KClients -benchmem -benchtime=5s > new.txt
$ benchcmp old.txt new.txt
benchmark                                       old ns/op     new ns/op     delta
BenchmarkServerGet1ReqPerConn10KClients-4       1031          1266          +22.79%
BenchmarkServerGet2ReqPerConn10KClients-4       771           997           +29.31%
BenchmarkServerGet10ReqPerConn10KClients-4      488           662           +35.66%
BenchmarkServerGet100ReqPerConn10KClients-4     460           601           +30.65%

benchmark                                       old allocs     new allocs     delta
BenchmarkServerGet1ReqPerConn10KClients-4       0              1              +Inf%
BenchmarkServerGet2ReqPerConn10KClients-4       0              1              +Inf%
BenchmarkServerGet10ReqPerConn10KClients-4      0              1              +Inf%
BenchmarkServerGet100ReqPerConn10KClients-4     0              1              +Inf%

benchmark                                       old bytes     new bytes     delta
BenchmarkServerGet1ReqPerConn10KClients-4       0             17            +Inf%
BenchmarkServerGet2ReqPerConn10KClients-4       0             17            +Inf%
BenchmarkServerGet10ReqPerConn10KClients-4      0             36            +Inf%
BenchmarkServerGet100ReqPerConn10KClients-4     4             27            +575.00%
```

Could you compare fasthttp benchmark results on your computer for the origin branch and for your branch and report results here? Benchmarks can be run with `go test -bench=. -benchmem -benchtime=5s`