@sprstnd 2016-01-06
the assembly produced by gc has less instructions and a simple benchmark of just the two methods(the old vs the new) shows improvement by 20%. Both methods produce a []byte and fill it. gc is optimized for conversations.

Im sorry, but I feel your benchmarks are flawed. Accept the changes or don't, doesn't bother me a bit. just thought I'd share some insight.